test-requirements.txt
flake8
httpretty
pytest
tox
tox-travis
twine
statsdmock